refactor(swarm): extract BackendRuntime trait + impls into runtime.rs
- Move BackendRuntime trait, SpacetimeRuntime, and ArcaneRuntime from
  main.rs to new runtime.rs module (349 lines)
- Extract control-mode orchestrator (run_control_mode + handle_control_connection)
  into control.rs (366 lines)
- Introduce SharedHandles struct to reduce repetitive .clone() calls
  inside each spawn_player implementation
- Update spawn_context.rs to use crate::runtime::BackendRuntime and
  SharedHandles instead of direct PlayerLoopShared passing
- Drop main.rs from 713 lines to 189 lines (under 350 target)
- Add 7 unit tests for the runtime module covering name(), spawn_read(),
  snapshot_cache_counters(), and SharedHandles construction

Closes #38
